+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - styling (used div-based tables, font-styling modified, show/hide exif-data by click)
 - added Camera Manufacturer
+- added credits
 
 ## 1.0.0 - 2016-05-28